Add initial structure of track-choosing website
So of course my initial concern is what javascript functions I will need and what their arguments should be
Secondarily I'm considering what the UI should be like. The jquery-ui will be vital for doing neat animations
With the idea of 'form follows function' we have the idea that the user will be answering questions, in a more-or-less sequential manner
Previous questions should be able to be revisited, but we assume this will not happen most of the time
Especially since it's not anticipated that anyone will use this for real.
Generally, the plan is to do the javascript, then whatever HTML is required will fall out of that, and any time remaining can be spent polishing animations